services/backanim/requirements.txt
requests==2.31.0
Pillow==10.2.0  # For image processing
numpy==1.26.4  # For image enhancement 